% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_func.R
\name{linearRegression}
\alias{linearRegression}
\title{Fitting Linear Models}
\usage{
linearRegression(data, predictors, response, intercept = TRUE)
}
\arguments{
\item{data}{The data set for multiple linear regression. It should be in the form of a 2d matrix}

\item{predictors}{A string array of all the predictors' names.}

\item{response}{A single string of the response.}

\item{intercept}{Boolean. TRUE by default for a model with intercept. Setting as FALSE will force the model to have no intercept.}
}
\value{
A list with 6 elements:
1. beta_hat; 2.R^2; 3.X; 4.Fitted values; 5.Residuals; 6.intercept or not
}
\description{
linearRegression is a mimic of lm() that is used to fit linear models.
}
\examples{
mt <- mtcars
linearRegression(mt, "mpg", "cyl", TRUE)

}
