% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_func.R
\name{model_residuals}
\alias{model_residuals}
\title{Summarize Model Residuals}
\usage{
model_residuals(model)
}
\arguments{
\item{model}{A list object returned by linearRegression.}
}
\value{
A table contains the regular residual, the standartized residual,
and the internally studentized residual of the model.
}
\description{
model_residuals is a function which extracts model residuals, standardized
residuals, and internally studentized residuals from the lists returned by
modeling functions
}
\examples{
mt <- mtcars
model_residuals(linearRegression(mt, "mpg", "cyl", TRUE))
}
